Refactor DateUtilsTest
